Refactor with openspec
