+++ app/src/main/java/com/finoldigital/ygolp/presentation/constants/Constants.kt
@@ -1,0 +1,5 @@
+package com.finoldigital.ygolp.presentation.constants
+
+const val MIN_LIFE_POINTS = 0
+const val MAX_LIFE_POINTS = 99999
+const val STARTING_LIFE_POINTS = 8000
