+++ app/src/main/java/com/finoldigital/ygolp/presentation/screens/LifePointsScreen.kt
@@ -1,120 +1,65 @@
 package com.finoldigital.ygolp.presentation.screens
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.detectHorizontalDragGestures
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.PagerState
+import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.VolumeOff
 import androidx.compose.material.icons.automirrored.filled.VolumeUp
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.wear.compose.material.Icon
 import androidx.wear.compose.material.MaterialTheme
-import androidx.wear.compose.material.Text
 import androidx.wear.compose.ui.tooling.preview.WearPreviewDevices
 import androidx.wear.compose.ui.tooling.preview.WearPreviewFontScales
 import com.finoldigital.ygolp.R
+import com.finoldigital.ygolp.presentation.components.LifePointsPage
 import com.finoldigital.ygolp.presentation.components.PlayerIndicator
+import com.finoldigital.ygolp.presentation.constants.STARTING_LIFE_POINTS
 import com.finoldigital.ygolp.presentation.enums.CalculatorMode
 import com.finoldigital.ygolp.presentation.enums.Player
 import com.finoldigital.ygolp.presentation.theme.AppColors
 
-const val STARTING_LIFE_POINTS = 8000
-
 @Composable
 fun LifePointsScreen(
-    player: Player = Player.ONE,
-    lifePoints: Int = STARTING_LIFE_POINTS,
+    pagerState: PagerState,
+    displayedLifePoints1: Int = STARTING_LIFE_POINTS,
+    displayedLifePoints2: Int = STARTING_LIFE_POINTS,
     isMuted: Boolean = false,
     onToggleMute: () -> Unit = {},
-    onShowCalculatorWithMode: (CalculatorMode) -> Unit = {},
-    onRestart: (() -> Unit)? = null,
+    onShowCalculatorWithMode: (Player, CalculatorMode) -> Unit = { _, _ -> },
+    onRestart: () -> Unit = {},
 ) {
-    val isLost = lifePoints <= MIN_LIFE_POINTS && onRestart != null
     MaterialTheme {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .then(
-                    if (isLost) {
-                        Modifier.clickable { onRestart() }
-                    } else {
-                        Modifier
-                    }
-                )
-        ) {
-            if (player == Player.ONE) {
-                Image(
-                    painterResource(R.drawable.lifepoints_background),
-                    contentDescription = stringResource(R.string.bg_player_1),
-                    modifier = Modifier.fillMaxSize(),
-                    contentScale = ContentScale.FillBounds
-                )
-            } else {
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .background(
-                            brush = Brush.verticalGradient(
-                                colors = listOf(
-                                    AppColors.P2GradientStart,
-                                    AppColors.P2GradientEnd
-                                )
-                            )
-                        )
+        Box(modifier = Modifier.fillMaxSize()) {
+            HorizontalPager(state = pagerState) { page ->
+                val player = if (page == 0) Player.ONE else Player.TWO
+                val displayedLifePoints = if (player == Player.ONE) displayedLifePoints1 else displayedLifePoints2
+                LifePointsPage(
+                    player = player,
+                    lifePoints = displayedLifePoints,
+                    onShowCalculatorWithMode = { calculatorMode ->
+                        onShowCalculatorWithMode(player, calculatorMode)
+                    },
+                    onRestart = if (displayedLifePoints <= 0) onRestart else null
                 )
             }
-
-            if (!isLost) {
-                Column(modifier = Modifier.fillMaxSize()) {
-                    Box(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxWidth()
-                            .clickable { onShowCalculatorWithMode(CalculatorMode.ADD) }
-                    )
-                    Box(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxWidth()
-                            .clickable { onShowCalculatorWithMode(CalculatorMode.SET) }
-                    )
-                    Box(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxWidth()
-                            .clickable { onShowCalculatorWithMode(CalculatorMode.SUBTRACT) }
-                    )
-                }
-            }
-            LifePointsText(lifePoints)
 
             PlayerIndicator(
                 modifier = Modifier
                     .align(Alignment.BottomCenter)
                     .padding(bottom = 16.dp),
-                player = player
+                player = if (pagerState.currentPage == 0) Player.ONE else Player.TWO
             )
 
-            // Mute/Unmute toggle button at top center
             Box(
                 modifier = Modifier
                     .align(Alignment.TopCenter)
@@ -134,35 +79,10 @@ fun LifePointsScreen(
     }
 }
 
-@Composable
-fun LifePointsText(lifePoints: Int) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        val lifePointsText =
-            if (lifePoints > MIN_LIFE_POINTS) lifePoints.toString() else stringResource(R.string.game_lost)
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            textAlign = TextAlign.Center,
-            color = AppColors.LpTextYellow,
-            fontFamily = FontFamily(Font(R.font.nationalyze_alp)),
-            fontSize = 32.sp,
-            text = lifePointsText
-        )
-    }
-}
-
 @WearPreviewDevices
 @WearPreviewFontScales
 @Composable
 fun LifePointsScreenPreview() {
-    LifePointsScreen()
-}
-
-@WearPreviewDevices
-@WearPreviewFontScales
-@Composable
-fun LifePointsScreenPreview2() {
-    LifePointsScreen(Player.TWO)
+    val pagerState = rememberPagerState(pageCount = { 2 })
+    LifePointsScreen(pagerState)
 }
