+++ openspec/changes/archive/2026-05-10-extract-lifepoints-horizontal-pager/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-10
