+++ openspec/changes/archive/2026-05-10-reorganize-presentation-packages/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-10
